Bank Marketing Data Classification
This report shows sequentially the application of a set of Data Science techniques to gain insights from the Direct Marketing campaign of a Portuguese Banking Institution. The data set has 45,212 instances and 20 features (input variables), which include the respondents’ level of education, social status, month and day, results of a previous marketing campaign, and some macroeconomic indicators.